Simplify bordered box margins
Now making use of background-clip: padding-box to prevent the background from filling in behind transparent borders, plus setting a transparent border to include the image within the box sizing.

This means the border falls within the sizing outlines, so margins actually represent the visible spacing from adjacent elements and not the bordered region as well! Yay!