Move DW locations off full index into DW indices
If we start adding every location to the main index it's gonna get long as hell. Switching to the same thing I was doing with the Light World locations, which is giving the LW folder a dedicated index page with the full set of links, and on the main index, just linking to that page's sections.

Also:
- made some minor tweaks to IDs ("queensmansion" to "queenmansion", to match things like "krisroom" and "noelleroom")
- changed "first floor" etc. in Ralsei's castle to "1F" etc. to match the in-game wording