Load special fonts on-demand
It's gonna slow pageloads down a lot if we're loading a bunch of special fonts Literally All The Time. When was the last time fnt_noelletiny was used

I guess an exception could be made for Dotum but I decided to put it in with the on-demand fonts anyway, because even if the whole Encounters category uses it, that's only 1 category out of 4. Also, dialogue isn't usually the very first thing on the page, so FOUS is going to be a rare thing.