Fix URL joining for nested list items
Holy shit I am so sick of this damn include. Okay I still don't understand what the problems with the linking were exactly but something about the way parents and mysplit were declared was getting shared across calls of this include?? So I rewrote that whole damn part with an array that tracks the pieces of the current path instead.